+++ TriangularLake/app/src/main/java/com/triangularlake/constantine/triangularlake/pojo/FavouriteProblemDTO.java
@@ -5,12 +5,12 @@ public class FavouriteProblemDTO {
     // Может и не правильно. Надо подумать.
     // + id поставил для адаптера, что бы удобно было строить view
 
-    public Long id;          // id проблемы
+    public Integer id;          // id проблемы
     public String name;      // название проблемы
     public String nameRu;    // название на русском проблемы
     public String grade;     // категория выбранной проблемы
 
-    public FavouriteProblemDTO(Long id, String name, String nameRu, String grade) {
+    public FavouriteProblemDTO(Integer id, String name, String nameRu, String grade) {
         this.id = id;
         this.name = name;
         this.nameRu = nameRu;
